backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib==21.2.0
